+++ src/platform/windows/events_loop.rs
@@ -12,41 +12,56 @@
 //! The closure passed to the `execute_in_thread` method takes an `Inserter` that you can use to
 //! add a `WindowState` entry to a list of window to be used by the callback.
 
+use std::{mem, ptr, thread};
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::ffi::OsString;
-use std::mem;
 use std::os::windows::ffi::OsStringExt;
 use std::os::windows::io::AsRawHandle;
-use std::ptr;
-use std::sync::mpsc;
-use std::sync::Arc;
-use std::sync::Barrier;
-use std::sync::Mutex;
-use std::sync::Condvar;
-use std::thread;
+use std::sync::{Arc, Barrier, Condvar, mpsc, Mutex};
 
-use winapi::shared::minwindef::{LOWORD, HIWORD, DWORD, WPARAM, LPARAM, INT, UINT, LRESULT, MAX_PATH};
+use winapi::ctypes::c_int;
+use winapi::shared::minwindef::{
+    BOOL,
+    DWORD,
+    HIWORD,
+    INT,
+    LOWORD,
+    LPARAM,
+    LRESULT,
+    MAX_PATH,
+    UINT,
+    WPARAM,
+};
 use winapi::shared::windef::{HWND, POINT, RECT};
 use winapi::shared::windowsx;
 use winapi::um::{winuser, shellapi, processthreadsapi};
-use winapi::um::winnt::{LONG, SHORT};
+use winapi::um::winnt::{LONG, LPCSTR, SHORT};
 
-use events::DeviceEvent;
+use {
+    ControlFlow,
+    CursorState,
+    Event,
+    EventsLoopClosed,
+    KeyboardInput,
+    LogicalCoordinates,
+    LogicalDimensions,
+    PhysicalDimensions,
+    WindowAttributes,
+    WindowEvent,
+    WindowId as SuperWindowId,
+};
+use events::{DeviceEvent, Touch, TouchPhase};
 use platform::platform::{event, Cursor, WindowId, DEVICE_ID, wrap_device_id, util};
+use platform::platform::dpi::{
+    become_dpi_aware,
+    dpi_to_scale_factor,
+    enable_non_client_dpi_scaling,
+    get_hwnd_scale_factor,
+};
 use platform::platform::event::{handle_extended_keys, process_key_params, vkey_to_winit_vkey};
-use platform::platform::raw_input::*;
+use platform::platform::raw_input::{get_raw_input_data, get_raw_mouse_button_state};
 use platform::platform::window::adjust_size;
-
-use ControlFlow;
-use CursorState;
-use Event;
-use EventsLoopClosed;
-use KeyboardInput;
-use WindowAttributes;
-use WindowEvent;
-use WindowId as SuperWindowId;
-use events::{Touch, TouchPhase};
 
 /// Contains saved window info for switching between fullscreen
 #[derive(Clone)]
@@ -98,11 +113,17 @@ pub struct EventsLoop {
     // Variable that contains the block state of the win32 event loop thread during a WM_SIZE event.
     // The mutex's value is `true` when it's blocked, and should be set to false when it's done
     // blocking. That's done by the parent thread when it receives a Resized event.
-    win32_block_loop: Arc<(Mutex<bool>, Condvar)>
+    win32_block_loop: Arc<(Mutex<bool>, Condvar)>,
 }
 
 impl EventsLoop {
     pub fn new() -> EventsLoop {
+        Self::with_dpi_awareness(true)
+    }
+
+    pub fn with_dpi_awareness(dpi_aware: bool) -> EventsLoop {
+        become_dpi_aware(dpi_aware);
+
         // The main events transfer channel.
         let (tx, rx) = mpsc::channel();
         let win32_block_loop = Arc::new((Mutex::new(false), Condvar::new()));
@@ -171,7 +192,7 @@ impl EventsLoop {
         EventsLoop {
             thread_id,
             receiver: rx,
-            win32_block_loop
+            win32_block_loop,
         }
     }
 
@@ -289,22 +310,21 @@ impl EventsLoopProxy {
     where
         F: FnMut(Inserter) + Send + 'static,
     {
-        unsafe {
-            // We are using double-boxing here because it make casting back much easier
-            let boxed = Box::new(function) as Box<FnMut(_)>;
-            let boxed2 = Box::new(boxed);
-            let raw = Box::into_raw(boxed2);
+        // We are using double-boxing here because it make casting back much easier
+        let double_box = Box::new(Box::new(function) as Box<FnMut(_)>);
+        let raw = Box::into_raw(double_box);
 
-            let res = winuser::PostThreadMessageA(
+        let res = unsafe {
+            winuser::PostThreadMessageA(
                 self.thread_id,
                 *EXEC_MSG_ID,
                 raw as *mut () as usize as WPARAM,
                 0,
-            );
-            // PostThreadMessage can only fail if the thread ID is invalid (which shouldn't happen
-            // as the events loop is still alive) or if the queue is full.
-            assert!(res != 0, "PostThreadMessage failed; is the messages queue full?");
-        }
+            )
+        };
+        // PostThreadMessage can only fail if the thread ID is invalid (which shouldn't happen as
+        // the events loop is still alive) or if the queue is full.
+        assert!(res != 0, "PostThreadMessage failed; is the messages queue full?");
     }
 }
 
@@ -313,7 +333,7 @@ lazy_static! {
     // WPARAM and LPARAM are unused.
     static ref WAKEUP_MSG_ID: u32 = {
         unsafe {
-            winuser::RegisterWindowMessageA("Winit::WakeupMsg\0".as_ptr() as *const i8)
+            winuser::RegisterWindowMessageA("Winit::WakeupMsg\0".as_ptr() as LPCSTR)
         }
     };
     // Message sent when we want to execute a closure in the thread.
@@ -321,14 +341,21 @@ lazy_static! {
     // and LPARAM is unused.
     static ref EXEC_MSG_ID: u32 = {
         unsafe {
-            winuser::RegisterWindowMessageA("Winit::ExecMsg\0".as_ptr() as *const i8)
+            winuser::RegisterWindowMessageA("Winit::ExecMsg\0".as_ptr() as LPCSTR)
         }
     };
     // Message sent by a `Window` when it wants to be destroyed by the main thread.
     // WPARAM and LPARAM are unused.
     pub static ref DESTROY_MSG_ID: u32 = {
         unsafe {
-            winuser::RegisterWindowMessageA("Winit::DestroyMsg\0".as_ptr() as *const i8)
+            winuser::RegisterWindowMessageA("Winit::DestroyMsg\0".as_ptr() as LPCSTR)
+        }
+    };
+    // Message sent by a `Window` after creation if it has a DPI != 96.
+    // WPARAM is the the DPI (u32). LOWORD of LPARAM is width, and HIWORD is height.
+    pub static ref INITIAL_DPI_MSG_ID: u32 = {
+        unsafe {
+            winuser::RegisterWindowMessageA("Winit::InitialDpiMsg\0".as_ptr() as LPCSTR)
         }
     };
 }
@@ -385,11 +412,18 @@ unsafe fn release_mouse() {
 //
 // Returning 0 tells the Win32 API that the message has been processed.
 // FIXME: detect WM_DWMCOMPOSITIONCHANGED and call DwmEnableBlurBehindWindow if necessary
-pub unsafe extern "system" fn callback(window: HWND, msg: UINT,
-                                       wparam: WPARAM, lparam: LPARAM)
-                                       -> LRESULT
-{
+pub unsafe extern "system" fn callback(
+    window: HWND,
+    msg: UINT,
+    wparam: WPARAM,
+    lparam: LPARAM,
+) -> LRESULT {
     match msg {
+        winuser::WM_NCCREATE => {
+            enable_non_client_dpi_scaling(window);
+            winuser::DefWindowProcW(window, msg, wparam, lparam)
+        },
+
         winuser::WM_CLOSE => {
             use events::WindowEvent::CloseRequested;
             send_event(Event::WindowEvent {
@@ -427,9 +461,14 @@ pub unsafe extern "system" fn callback(window: HWND, msg: UINT,
 
             let windowpos = lparam as *const winuser::WINDOWPOS;
             if (*windowpos).flags & winuser::SWP_NOMOVE != winuser::SWP_NOMOVE {
+                let dpi_factor = get_hwnd_scale_factor(window);
+                let logical_position = LogicalCoordinates::from_physical(
+                    ((*windowpos).x, (*windowpos).y),
+                    dpi_factor,
+                );
                 send_event(Event::WindowEvent {
                     window_id: SuperWindowId(WindowId(window)),
-                    event: Moved((*windowpos).x, (*windowpos).y),
+                    event: Moved(logical_position),
                 });
             }
 
@@ -448,9 +487,11 @@ pub unsafe extern "system" fn callback(window: HWND, msg: UINT,
                 let mut context_stash = context_stash.borrow_mut();
                 let cstash = context_stash.as_mut().unwrap();
 
+                let dpi_factor = get_hwnd_scale_factor(window);
+                let logical_size = LogicalDimensions::from_physical((w, h), dpi_factor);
                 let event = Event::WindowEvent {
                     window_id: SuperWindowId(WindowId(window)),
-                    event: Resized(w, h),
+                    event: Resized(logical_size),
                 };
 
                 // If this window has been inserted into the window map, the resize event happened
@@ -1007,9 +1048,81 @@ pub unsafe extern "system" fn callback(window: HWND, msg: UINT,
             0
         },
 
+        // Only sent on Windows 8.1 or newer. On Windows 7 and older user has to log out to change
+        // DPI, therefore all applications are closed while DPI is changing.
+        winuser::WM_DPICHANGED => {
+            use events::WindowEvent::HiDpiFactorChanged;
+
+            // This message actually provides two DPI values - x and y. However MSDN says that
+            // "you only need to use either the X-axis or the Y-axis value when scaling your
+            // application since they are the same".
+            // https://msdn.microsoft.com/en-us/library/windows/desktop/dn312083(v=vs.85).aspx
+            let dpi_x = u32::from(LOWORD(wparam as DWORD));
+
+            // Resize window to the size suggested by Windows.
+            let rect = &*(lparam as *const RECT);
+            winuser::SetWindowPos(
+                window,
+                ptr::null_mut(),
+                rect.left,
+                rect.top,
+                rect.right - rect.left,
+                rect.bottom - rect.top,
+                winuser::SWP_NOZORDER | winuser::SWP_NOACTIVATE,
+            );
+
+            send_event(Event::WindowEvent {
+                window_id: SuperWindowId(WindowId(window)),
+                event: HiDpiFactorChanged(dpi_to_scale_factor(dpi_x)),
+            });
+
+            0
+        },
+
         _ => {
             if msg == *DESTROY_MSG_ID {
                 winuser::DestroyWindow(window);
+                0
+            } else if msg == *INITIAL_DPI_MSG_ID {
+                use events::WindowEvent::HiDpiFactorChanged;
+                let scale_factor = dpi_to_scale_factor(wparam as u32);
+                send_event(Event::WindowEvent {
+                    window_id: SuperWindowId(WindowId(window)),
+                    event: HiDpiFactorChanged(scale_factor),
+                });
+                // Automatically resize for actual DPI
+                let width = LOWORD(lparam as DWORD) as u32;
+                let height = HIWORD(lparam as DWORD) as u32;
+                let (adjusted_width, adjusted_height): (u32, u32) = PhysicalDimensions::from_logical(
+                    (width, height),
+                    scale_factor,
+                ).into();
+                // We're not done yet! `SetWindowPos` needs the window size, not the client area size.
+                let mut rect = RECT {
+                    top: 0,
+                    left: 0,
+                    bottom: adjusted_height as LONG,
+                    right: adjusted_width as LONG,
+                };
+                let dw_style = winuser::GetWindowLongA(window, winuser::GWL_STYLE) as DWORD;
+                let b_menu = !winuser::GetMenu(window).is_null() as BOOL;
+                let dw_style_ex = winuser::GetWindowLongA(window, winuser::GWL_EXSTYLE) as DWORD;
+                winuser::AdjustWindowRectEx(&mut rect, dw_style, b_menu, dw_style_ex);
+                let outer_x = (rect.right - rect.left).abs() as c_int;
+                let outer_y = (rect.top - rect.bottom).abs() as c_int;
+                winuser::SetWindowPos(
+                    window,
+                    ptr::null_mut(),
+                    0,
+                    0,
+                    outer_x,
+                    outer_y,
+                    winuser::SWP_NOMOVE
+                    | winuser::SWP_NOREPOSITION
+                    | winuser::SWP_NOZORDER
+                    | winuser::SWP_NOACTIVATE,
+                );
+                // TODO: Adjust min+max size
                 0
             } else {
                 winuser::DefWindowProcW(window, msg, wparam, lparam)