+++ src/platform/windows/monitor.rs
@@ -1,38 +1,30 @@
-use winapi::ctypes::wchar_t;
-use winapi::shared::minwindef::{DWORD, LPARAM, BOOL, TRUE};
-use winapi::shared::windef::{HMONITOR, HDC, LPRECT, HWND};
+use winapi::shared::minwindef::{BOOL, DWORD, LPARAM, TRUE};
+use winapi::shared::windef::{HDC, HMONITOR, HWND, LPRECT, POINT};
 use winapi::um::winuser;
 
-use std::collections::VecDeque;
 use std::{mem, ptr};
+use std::collections::VecDeque;
 
 use super::{EventsLoop, util};
+use platform::platform::dpi::{dpi_to_scale_factor, get_monitor_dpi};
 
 /// Win32 implementation of the main `MonitorId` object.
 #[derive(Debug, Clone)]
 pub struct MonitorId {
-    /// The system name of the adapter.
-    adapter_name: [wchar_t; 32],
-
     /// Monitor handle.
     hmonitor: HMonitor,
-
     /// The system name of the monitor.
     monitor_name: String,
-
     /// True if this is the primary monitor.
     primary: bool,
-
     /// The position of the monitor in pixels on the desktop.
     ///
     /// A window that is positioned at these coordinates will overlap the monitor.
     position: (i32, i32),
-
     /// The current resolution in pixels on the monitor.
     dimensions: (u32, u32),
-
-    /// DPI scaling factor.
-    hidpi_factor: f32,
+    /// DPI scale factor.
+    hidpi_factor: f64,
 }
 
 // Send is not implemented for HMONITOR, we have to wrap it and implement it manually.
@@ -44,122 +36,109 @@ struct HMonitor(HMONITOR);
 
 unsafe impl Send for HMonitor {}
 
-unsafe extern "system" fn monitor_enum_proc(hmonitor: HMONITOR, _: HDC, place: LPRECT, data: LPARAM) -> BOOL {
+unsafe extern "system" fn monitor_enum_proc(
+    hmonitor: HMONITOR,
+    _hdc: HDC,
+    _place: LPRECT,
+    data: LPARAM,
+) -> BOOL {
     let monitors = data as *mut VecDeque<MonitorId>;
-
-    let place = *place;
-    let position = (place.left as i32, place.top as i32);
-    let dimensions = ((place.right - place.left) as u32, (place.bottom - place.top) as u32);
-
-    let mut monitor_info: winuser::MONITORINFOEXW = mem::zeroed();
-    monitor_info.cbSize = mem::size_of::<winuser::MONITORINFOEXW>() as DWORD;
-    if winuser::GetMonitorInfoW(hmonitor, &mut monitor_info as *mut winuser::MONITORINFOEXW as *mut winuser::MONITORINFO) == 0 {
-        // Some error occurred, just skip this monitor and go on.
-        return TRUE;
-    }
-
-    (*monitors).push_back(MonitorId {
-        adapter_name: monitor_info.szDevice,
-        hmonitor: HMonitor(hmonitor),
-        monitor_name: util::wchar_to_string(&monitor_info.szDevice),
-        primary: monitor_info.dwFlags & winuser::MONITORINFOF_PRIMARY != 0,
-        position,
-        dimensions,
-        hidpi_factor: 1.0,
-    });
-
-    // TRUE means continue enumeration.
-    TRUE
+    (*monitors).push_back(MonitorId::from_hmonitor(hmonitor));
+    TRUE // continue enumeration
 }
 
 impl EventsLoop {
+    // TODO: Investigate opportunities for caching
     pub fn get_available_monitors(&self) -> VecDeque<MonitorId> {
+        let mut monitors: VecDeque<MonitorId> = VecDeque::new();
         unsafe {
-            let mut result: VecDeque<MonitorId> = VecDeque::new();
-            winuser::EnumDisplayMonitors(ptr::null_mut(), ptr::null_mut(), Some(monitor_enum_proc), &mut result as *mut _ as LPARAM);
-            result
+            winuser::EnumDisplayMonitors(
+                ptr::null_mut(),
+                ptr::null_mut(),
+                Some(monitor_enum_proc),
+                &mut monitors as *mut _ as LPARAM,
+            );
         }
+        monitors
     }
 
-    pub fn get_current_monitor(handle: HWND) -> MonitorId {
-        unsafe {
-            let mut monitor_info: winuser::MONITORINFOEXW = mem::zeroed();
-            monitor_info.cbSize = mem::size_of::<winuser::MONITORINFOEXW>() as DWORD;
-
-            let hmonitor = winuser::MonitorFromWindow(handle, winuser::MONITOR_DEFAULTTONEAREST);
-
-            winuser::GetMonitorInfoW(
-                hmonitor,
-                &mut monitor_info as *mut winuser::MONITORINFOEXW as *mut winuser::MONITORINFO,
-            );
-
-            let place = monitor_info.rcMonitor;
-            let position = (place.left as i32, place.top as i32);
-            let dimensions = (
-                (place.right - place.left) as u32,
-                (place.bottom - place.top) as u32,
-            );
-
-            MonitorId {
-                adapter_name: monitor_info.szDevice,
-                hmonitor: super::monitor::HMonitor(hmonitor),
-                monitor_name: util::wchar_to_string(&monitor_info.szDevice),
-                primary: monitor_info.dwFlags & winuser::MONITORINFOF_PRIMARY != 0,
-                position,
-                dimensions,
-                hidpi_factor: 1.0,
-            }
-        }
+    pub fn get_current_monitor(hwnd: HWND) -> MonitorId {
+        let hmonitor = unsafe {
+            winuser::MonitorFromWindow(hwnd, winuser::MONITOR_DEFAULTTONEAREST)
+        };
+        MonitorId::from_hmonitor(hmonitor)
     }
 
     pub fn get_primary_monitor(&self) -> MonitorId {
-        // we simply get all available monitors and return the one with the `MONITORINFOF_PRIMARY` flag
-        // TODO: it is possible to query the win32 API for the primary monitor, this should be done
-        //  instead
-        for monitor in self.get_available_monitors().into_iter() {
-            if monitor.primary {
-                return monitor;
-            }
-        }
+        const ORIGIN: POINT = POINT { x: 0, y: 0 };
+        let hmonitor = unsafe {
+            winuser::MonitorFromPoint(ORIGIN, winuser::MONITOR_DEFAULTTOPRIMARY)
+        };
+        MonitorId::from_hmonitor(hmonitor)
+    }
+}
 
-        panic!("Failed to find the primary monitor")
+fn get_monitor_info(hmonitor: HMONITOR) -> Result<winuser::MONITORINFOEXW, util::WinError> {
+    let mut monitor_info: winuser::MONITORINFOEXW = unsafe { mem::uninitialized() };
+    monitor_info.cbSize = mem::size_of::<winuser::MONITORINFOEXW>() as DWORD;
+    let status = unsafe {
+        winuser::GetMonitorInfoW(
+            hmonitor,
+            &mut monitor_info as *mut winuser::MONITORINFOEXW as *mut winuser::MONITORINFO,
+        )
+    };
+    if status == 0 {
+        Err(util::WinError::from_last_error())
+    } else {
+        Ok(monitor_info)
     }
 }
 
 impl MonitorId {
-    /// See the docs if the crate root file.
+    pub(crate) fn from_hmonitor(hmonitor: HMONITOR) -> Self {
+        let monitor_info = get_monitor_info(hmonitor).expect("`GetMonitorInfoW` failed");
+        let place = monitor_info.rcMonitor;
+        let dimensions = (
+            (place.right - place.left) as u32,
+            (place.bottom - place.top) as u32,
+        );
+        MonitorId {
+            hmonitor: HMonitor(hmonitor),
+            monitor_name: util::wchar_ptr_to_string(monitor_info.szDevice.as_ptr()),
+            primary: util::has_flag(monitor_info.dwFlags, winuser::MONITORINFOF_PRIMARY),
+            position: (place.left as i32, place.top as i32),
+            dimensions,
+            hidpi_factor: dpi_to_scale_factor(get_monitor_dpi(hmonitor).unwrap_or(96)),
+        }
+    }
+
     #[inline]
     pub fn get_name(&self) -> Option<String> {
         Some(self.monitor_name.clone())
     }
 
-    /// See the docs of the crate root file.
     #[inline]
     pub fn get_native_identifier(&self) -> String {
         self.monitor_name.clone()
     }
 
-    /// See the docs of the crate root file.
     #[inline]
     pub fn get_hmonitor(&self) -> HMONITOR {
         self.hmonitor.0
     }
 
-    /// See the docs of the crate root file.
     #[inline]
     pub fn get_dimensions(&self) -> (u32, u32) {
-        // TODO: retrieve the dimensions every time this is called
         self.dimensions
     }
 
-    /// A window that is positioned at these coordinates will overlap the monitor.
     #[inline]
     pub fn get_position(&self) -> (i32, i32) {
         self.position
     }
 
     #[inline]
-    pub fn get_hidpi_factor(&self) -> f32 {
+    pub fn get_hidpi_factor(&self) -> f64 {
         self.hidpi_factor
     }
 }