+++ src/window.rs
@@ -1,7 +1,5 @@
 use std::collections::vec_deque::IntoIter as VecDequeIter;
 
-use libc;
-
 use {
     CreationError,
     CursorState,
@@ -148,6 +146,7 @@ impl WindowBuilder {
     ///
     /// Error should be very rare and only occur in case of permission denied, incompatible system,
     /// out of memory, etc.
+    #[inline]
     pub fn build(mut self, events_loop: &EventsLoop) -> Result<Window, CreationError> {
         self.window.dimensions = Some(self.window.dimensions.unwrap_or_else(|| {
             if let Some(ref monitor) = self.window.fullscreen {
@@ -309,26 +308,9 @@ impl Window {
         self.window.set_resizable(resizable)
     }
 
-    /// DEPRECATED. Gets the native platform specific display for this window.
-    /// This is typically only required when integrating with
-    /// other libraries that need this information.
-    #[deprecated]
-    #[inline]
-    pub unsafe fn platform_display(&self) -> *mut libc::c_void {
-        self.window.platform_display()
-    }
-
-    /// DEPRECATED. Gets the native platform specific window handle. This is
-    /// typically only required when integrating with other libraries
-    /// that need this information.
-    #[deprecated]
-    #[inline]
-    pub unsafe fn platform_window(&self) -> *mut libc::c_void {
-        self.window.platform_window()
-    }
-
     /// Modifies the mouse cursor of the window.
     /// Has no effect on Android.
+    #[inline]
     pub fn set_cursor(&self, cursor: MouseCursor) {
         self.window.set_cursor(cursor);
     }
@@ -403,6 +385,7 @@ impl Window {
     }
 
     /// Returns the monitor on which the window currently resides
+    #[inline]
     pub fn get_current_monitor(&self) -> MonitorId {
         self.window.get_current_monitor()
     }
@@ -468,7 +451,7 @@ impl MonitorId {
     /// On X11 the DPI factor can be overridden using the `WINIT_HIDPI_FACTOR` environment
     /// variable.
     #[inline]
-    pub fn get_hidpi_factor(&self) -> f32 {
+    pub fn get_hidpi_factor(&self) -> f64 {
         self.inner.get_hidpi_factor()
     }
 }